alpha_0.6.5

+++ scripts/configure_bridge.sh
@@ -9,7 +9,8 @@ fi
 bridge()
 {
     # Caminho para o arquivo de configuração
-    config_file="../configs/network.conf"
+    cd /Proxmox-Debian12
+    config_file="configs/network.conf"
 
     # Verificar se o arquivo de configuração existe
     if [ ! -f "$config_file" ]; then
@@ -28,10 +29,16 @@ bridge()
     echo "Gateway: $GATEWAY"
 
     # Utilizar as variáveis lidas do arquivo ou solicitar novas se estiverem em branco
-    read -p "Informe o nome da interface física (deixe em branco para manter $INTERFACE): " interface_fisica
-    read -p "Informe o endereço IP para a bridge (deixe em branco para manter $IP_ADDRESS): " endereco_ip
-    read -p "Informe a máscara de sub-rede para a bridge (deixe em branco para manter $MASCARA_SUBREDE): " mascara_subrede
-    read -p "Informe o gateway para a bridge (deixe em branco para manter $GATEWAY): " gateway
+    echo -e "Deseja editar as informações da interface de rede para criação da bridge?"
+    echo -e "([S]im para editar [N]ão para continuar com a interface listada acima)"
+    read -p "Resposta: " editar
+
+    if [[ "$editar" =~ ^[Ss](im)?$ ]]; then
+        read -p "Informe o nome da interface física (deixe em branco para manter $INTERFACE): " interface_fisica
+        read -p "Informe o endereço IP para a bridge (deixe em branco para manter $IP_ADDRESS): " endereco_ip
+        read -p "Informe a máscara de sub-rede para a bridge (deixe em branco para manter $MASCARA_SUBREDE): " mascara_subrede
+        read -p "Informe o gateway para a bridge (deixe em branco para manter $GATEWAY): " gateway
+    fi
 
     # Utilizar as variáveis lidas ou as novas informadas
     INTERFACE=${interface_fisica:-$INTERFACE}
@@ -40,8 +47,17 @@ bridge()
     GATEWAY=${gateway:-$GATEWAY}
 
 
-# Criar a bridge vmbr0
-echo "Criando a bridge vmbr0..."
+# Verificar se a bridge vmbr0 já existe
+if grep -q "iface vmbr0" /etc/network/interfaces; then
+    echo "A bridge vmbr0 já existe. Reconfigurando..."
+    
+    # Remover a configuração existente
+    sed -i '/iface vmbr0/,/^$/d' /etc/network/interfaces
+else
+    echo "Criando a bridge vmbr0..."
+fi
+
+# Criar a bridge vmbr0 com as novas informações
 cat <<EOF >> /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
@@ -58,12 +74,13 @@ EOF
     systemctl restart networking
 
     echo "A bridge vmbr0 foi criada com sucesso!"
+    cd /
 }
 
 reboot()
 {
     echo -e "\e[1;32mInstalação e configuração do Proxmox concluída com sucesso!\e[0m"
-    echo -e "\e[1;91mAVISO: O sistema será reiniciado automaticamente para concluir a instalação.\e[0m"
+    echo -e "\e[1;91mAVISO: O sistema será reiniciado.\e[0m"
     sleep 5
     systemctl reboot
 }
@@ -71,7 +88,14 @@ reboot()
 main()
 {
     bridge
-    reboot
+
+    echo -e "Deseja reiniciar o computador agora? (Opcional)"
+    read -p "Resposta " perguntar_reboot
+
+    if ["$perguntar_reboot" == "sim"]; then
+        reboot
+    fi
+    
 }
 
 main
